backend/requirements.txt
fastapi
pydantic
python-dotenv
mira-sdk
uvicorn
